Add RESET to the printing of debug lines
Add RESET to the printing of debug lines otherwise the last color set remains
in the terminal on program exit.

$Term::ANSIColor::AUTORESET would do this automatically if we didn't use commas
for the print statement, but because we are getting the color from a variable I
am unaware of the Perl voodoo required to make it work with a variable instead
of the Term::ANSIColor constant.  I have a feeling this is more readable
anyway.