Delay daemonization until last moment possible
Prior to this change, tenshi daemonized before it ran through initialization
steps.  This had the unfortunate consequence that, if there happened to be an
error while initializing, the caller would be unaware that tenshi failed to
actually start (the parent process had exited cleanly already).  This patch
moves the actual daemonization as close to the running loop as possible and
also makes some of the process preparations also apply to non-daemon modes (it
would be misleading to have the non-daemon modes succeed where the daemon mode
fails).