Use pipe instead of open with "|-" mode for tail
Using open with the "|-" or "-|" modes adds an implicit close/wait on exit.
This means that if we open the tail processes before daemonizing, the exiting
parent will hang and tenshi will appear broken.

This didn't appear to be an issue before because this behavior doesn't apply to
fds 0, 1, or 2.  Since we close STDIN/STDOUT/STDERR and the new "|-" pipes get
opened, they usually get opened as fds 0, 1, or 2, thus hiding this problem
unless multiple_tails is on and several files are configured.